feat(shopping,tasks,folders): decimal prices, receipts, task filters, folder rename/drag-drop
- Shopping: estimated/actual price and alternative price widened Int->Double
  end-to-end (model, Room migration 11->12, form parsing/keyboard, budget math,
  web-import mapper) so fractional amounts like 5804.25 no longer get mangled
  into 580425. List now sorts need -> ordered -> bought instead of insertion
  order. Items can attach receipts/documents (Room migration 12->13 adds
  shopping_item_id to documents), reusing the same attach/scan flow tasks use.
- Tasks: assignee form no longer offers "Unassigned" (must pick a partner or
  Both); added priority and assigned-to filter chips alongside the existing
  tag filter. Fixed inability to delete the last task in a short list - the
  FAB was sitting directly over the last row's delete button.
- Folders: documents can be renamed (same dialog pattern as folder rename)
  and dragged onto a folder to move them, with a follow-the-finger drag
  animation (offset/scale/elevation) and a spring-back on drop.
- Also bumped shopping/folders list bottom padding to keep the FAB clear of
  the last row.

+++ android/core/domain/src/main/kotlin/com/oryareach/core/domain/shopping/Budget.kt
@@ -4,11 +4,11 @@ import com.oryareach.core.model.ShoppingCategory
 import com.oryareach.core.model.ShoppingItem
 import com.oryareach.core.model.ShoppingStatus
 
-data class CategoryTotals(val category: ShoppingCategory, val estimated: Int, val spent: Int)
+data class CategoryTotals(val category: ShoppingCategory, val estimated: Double, val spent: Double)
 
 data class BudgetTotals(
-    val estimatedTotal: Int,
-    val spentTotal: Int,
+    val estimatedTotal: Double,
+    val spentTotal: Double,
     val boughtCount: Int,
     val totalCount: Int,
     val byCategory: List<CategoryTotals>,
@@ -18,7 +18,7 @@ data class BudgetTotals(
  * `actualPrice` wins over everything else — it is what was really paid. Absent that, a
  * chosen alternative's own price stands in; only then does the original estimate apply.
  */
-fun itemEffectivePrice(item: ShoppingItem): Int? {
+fun itemEffectivePrice(item: ShoppingItem): Double? {
     item.actualPrice?.let { return it }
     item.chosenAlternativeId?.let { chosenId ->
         item.alternatives.firstOrNull { it.id == chosenId }?.price?.let { return it }
@@ -27,19 +27,19 @@ fun itemEffectivePrice(item: ShoppingItem): Int? {
 }
 
 fun calculateBudget(items: List<ShoppingItem>): BudgetTotals {
-    val byCategory = linkedMapOf<ShoppingCategory, MutableList<Int>>()
-    var estimatedTotal = 0
-    var spentTotal = 0
+    val byCategory = linkedMapOf<ShoppingCategory, MutableList<Double>>()
+    var estimatedTotal = 0.0
+    var spentTotal = 0.0
     var boughtCount = 0
 
     for (item in items) {
-        val estimated = item.estimatedPrice ?: itemEffectivePrice(item) ?: 0
-        val spent = if (item.status == ShoppingStatus.BOUGHT) itemEffectivePrice(item) ?: 0 else 0
+        val estimated = item.estimatedPrice ?: itemEffectivePrice(item) ?: 0.0
+        val spent = if (item.status == ShoppingStatus.BOUGHT) itemEffectivePrice(item) ?: 0.0 else 0.0
         estimatedTotal += estimated
         spentTotal += spent
         if (item.status == ShoppingStatus.BOUGHT) boughtCount += 1
 
-        val bucket = byCategory.getOrPut(item.category) { mutableListOf(0, 0) }
+        val bucket = byCategory.getOrPut(item.category) { mutableListOf(0.0, 0.0) }
         bucket[0] += estimated
         bucket[1] += spent
     }
